+++ src/Core/Data/Dto/Identity/UpdateUserPermissionsRequestDto.cs
@@ -6,7 +6,7 @@ namespace GamaEdtech.Data.Dto.Identity
 
     public sealed class UpdateUserPermissionsRequestDto
     {
-        public int UserId { get; set; }
+        public long UserId { get; set; }
         public required IEnumerable<string?> Permissions { get; set; }
         public SystemClaim? SystemClaims { get; set; }
         public Role? Roles { get; set; }
