+++ src/Domain/Specification/Message/SenderIdEqualsSpecification.cs
@@ -5,7 +5,7 @@ namespace GamaEdtech.Domain.Specification.Message
     using GamaEdtech.Common.DataAccess.Specification;
     using GamaEdtech.Domain.Entity;
 
-    public sealed class SenderIdEqualsSpecification(int senderId) : SpecificationBase<Message>
+    public sealed class SenderIdEqualsSpecification(long senderId) : SpecificationBase<Message>
     {
         public override Expression<Func<Message, bool>> Expression() => (t) => t.SenderId == senderId;
     }
