+++ src/Presentation/Api/Areas/Admin/Controllers/PaymentsController.cs
@@ -41,7 +41,7 @@ namespace GamaEdtech.Presentation.Api.Areas.Admin.Controllers
 
                 if (request.UserId.HasValue)
                 {
-                    var spec = new UserIdEqualsSpecification<Payment, int>(request.UserId.Value);
+                    var spec = new UserIdEqualsSpecification<Payment, long>(request.UserId.Value);
                     specification = specification is null ? spec : specification.And(spec);
                 }
